refactor: remove fixed size string buffers
before, there were a lot of buffers of size DEFAULT_STR_SIZE used to
store e.g. identifiers. It used to be set to 32 bytes.

Now the identifier size is not limited anymore, and also no space will
be wasted in memory if the identifier is smaller. We only allocate as
much as is actually needed by the string.