compiler: avr: do not branch with limited range
Before, the code was branching using 'brne','brlt' and such directly to
the target label.

But the distance of that label is unknown at that time. And
brne,brle,breq,... all can only branch +/- 64 words from the instruction
pointer.

So they cannot be used and 'jmp' must be used. We can skip the 'jmp' if
needed by using 'brne','brlt','breq'..., inverting the original
condition.

Signed-off-by: Alexander Hansen <alex23667@gmail.com>