compiler: implement basic liveness
IR statements can 'use' and 'define' IR temporaries.

This can be used to calculate sets 'in', 'out'
which describe which temporaries are live on an incoming respective
outgoing edge of an IR statement.

2 IR temporaries (used by struct TAC in our case) can share a register
if they are not both live at the some point in the program.

Tests: 2 very basic tests, needs to be extended
      TODO: probably make a '_liveness_struct.h' file
      to define 'struct Liveness' so it can be inspected
      from the testcases.

References:
  https://lambda.uta.edu/cse5317/notes/node40.html