compiler: x86: callee save/restore registers
In x86-64 there are some callee-saved registers.
Write some function to determine which registers
the callee must save, depending on which registers it is
touching (inspect struct RAT).

To err on the side of caution, we can always save and restore more
registers than actually needed. It will just make the code bigger.

Stack frame layout:

-----
return address
-----
...
callee saved registers <- rbp
-----
local variables
-----
...
saved args <- rsp
-----
...
...      0x0

The registers are saved as part of TAC_SETUP_STACKFRAME and restored as
part of TAC_RETURN. This is not optimal since TAC_RETURN can appear
multiple times and the code to restore the registers will then be
duplicated. It would be great to have some kind of IR statement to
designate the function epilogue to restore the callers registers and
finally 'ret'. The TAC_RETURN could then just 'jmp' there.

References:
  https://www.agner.org/optimize/calling_conventions.pdf