examples: rework examples
Add assert_examples.sh

That script loops through all ${filename}.dg files in examples/
and if ${filename}.exitcode exists,
it compiles that example program and compares the exit code with whats
in ${filename}.exitcode.

If the exit code of the compiled example program matches, that test is
PASS, otherwise it is FAIL.

This makes for an easy way to write tests for various compiler features.