compiler: fix mdirect expr
previously, it used to be e.g.
```
	int8 x = 3;
	[0x100] = x;

	x = [0x32];
```
and the width of the load/store would be implicit.

But thats not safe against refactoring of the type of 'x' for example.

Also makes it hard to handle cases like
```
[0x100] = (0x38 + x);
```

where the width of the expression on the right-hand side is not known at
compile-time.

So, modify the mdirect expr to have a constant for the width of the
load/store.

```
[0x100, 1] = 0x38;
uint16 y = [0x200, 1];
```