TAC: refactor int32_t tac_max_temp(struct TAC* tac)
Previous, this function used to access 'tac' without any knowledge of
its kind.

Refactor it to return the correct value based on 'tac->kind'.

This is inconsequential for this function as the relevant
IR temporary indices should be 0-initialized anyways.

But this way it looks cooler and will cause an error at runtime if a
case is not implemented.