compiler: x86: add stdlib/syscalls.dg by default
To allow to use syscalls, compile stdlib/syscalls.dg by default. This
only goes into effect if

- sd is called from commandline (not from test)
- -x86 flag is passed

There is currently some workaround since the compiler is only supposed
to handle one '.tokens' file.

So the stdlib/syscalls.dg is not handled 'properly'.

Also, the compiler now knows about the absolute path of stdlib/, which
is built into 'sd' during compilation.

So this can be extended in the future to add files to be compiled
'by default', which can be used for stdlib.

For this to work properly and not cause too much overhead, the dead-code
analysis must work well to avoid emitting code for functions which are
never called!

Since the '@syscall' definitions do not cause any extra code-gen for
these functions, it will be fine for now.