lexer: do not create .tokens file by default
Refactor the lexer to work with file descriptors.
Use memfd_create to create an in-memory file descriptor.

There is still some overhead by serializing and deserializing the
tokens. Which can be removed in a future patch.

There is a new flag to dump the tokens to file:

sd -dump-tokens simple.dg