compiler: refactor anytype '#' handling
struct Type now has a property 'bool is_anytype' to represent the
anytype.

Anytype can be used to override the typechecker. For example, calling a
memory allocation function returns a pointer, which can be a used as a
pointer to any other type, whatever is put there.

Anytype should have the highest native width of the architecture which
is targeted, so 8 bytes for x86-64, and 2 bytes for AVR.