tac: add line number information
Before, TAC IR buffers lost the source location, making it difficult to
guess which IR statement was generated via which AST node and source
statement.

Now that information is present in 'struct TAC' for debug output.

```
0   L115  println_uint64:
1   L115  setup_stackframe 8
2         load t499 = &l0 (x) (8 bytes)
3   L117  t500 = [t499] (8 bytes)
4   L117  param t500 (index 0)
5   L117  t501 = call print_uint64
6   L118  t502 = call println
7   L120  t503 = 0
8   L120  return t503
```

The first number is the index of the IR statement in the buffer,
the second is the line number.

Sometimes the line number is missing, which can be due to DerefLL and
other intermediary abstractions.

resolves issue #163